delete old layout
